Split the implementation and interface in `miral::Decoration`
The implementation is now called `UserDecoration` and is implemented as
an external class in miral-shell. The interface does book keeping of the
necessary data members.
