+++ src/miral/decoration_basic_manager.cpp
@@ -15,25 +15,82 @@
  */
 
 #include "miral/decoration_basic_manager.h"
-#include "miral/decoration_manager_builder.h"
-
+#include "mir/scene/session.h"
+#include "mir/scene/surface.h"
+#include "mir/server.h"
 #include "mir/shell/decoration/basic_manager.h"
-#include "mir/shell/decoration.h"
-#include "decoration_adapter.h"
+#include "mir/shell/shell.h"
+#include "mir/shell/surface_specification.h"
+#include "mir/optional_value.h"
+#include "mir/wayland/weak.h"
+#include "mir/compositor/buffer_stream.h"
+
+#include "miral/decoration_adapter.h"
+#include "miral/decoration_manager_builder.h"
 
 #include <memory>
 #include <utility>
 
+auto create_surface(
+    std::shared_ptr<ms::Surface> window_surface,
+    std::shared_ptr<msh::Shell> shell) -> std::shared_ptr<ms::Surface>
+{
+    auto const& session = window_surface->session().lock();
+    msh::SurfaceSpecification params;
+    params.type = mir_window_type_decoration;
+    params.parent = window_surface;
+    auto const size = window_surface->window_size();
+    params.width = size.width;
+    params.height = size.height;
+    params.aux_rect = {{}, {}};
+    params.aux_rect_placement_gravity = mir_placement_gravity_northwest;
+    params.surface_placement_gravity = mir_placement_gravity_northwest;
+    params.placement_hints = MirPlacementHints(0);
+    // Will be replaced by initial update
+    params.streams = {{
+        session->create_buffer_stream(mg::BufferProperties{
+            geom::Size{1, 1},
+            buffer_format,
+            mg::BufferUsage::software}),
+        {},
+        }};
+    return shell->create_surface(session, {}, params, nullptr, nullptr);
+}
+
 struct miral::DecorationBasicManager::Self : public mir::shell::decoration::BasicManager
 {
-    Self(Decoration::DecorationBuilder decoration_builder) :
-        mir::shell::decoration::BasicManager(decoration_builder)
+    Self(mir::Server& server, Decoration::DecorationBuilder&& decoration_builder) :
+        mir::shell::decoration::BasicManager(
+            [&server,decoration_builder](auto shell, auto window_surface)
+            {
+                auto session = window_surface->session().lock();
+                auto decoration_surface = create_surface(window_surface, shell);
+
+                auto decoration_adapter = decoration_builder(shell, window_surface);
+                auto decoration = decoration_adapter->to_decoration();
+
+                decoration_adapter->init(window_surface, decoration_surface, server.the_buffer_allocator(), shell);
+
+                // Rendering events come after this point
+                decoration_adapter->redraw_notifier()->register_listener(
+                    [decoration_adapter, decoration]()
+                    {
+                        decoration_adapter->update();
+                    });
+
+                // Initial redraw to set up the margins and buffers for decorations.
+                // Just like BasicDecoration
+                decoration_adapter->redraw_notifier()->notify();
+
+                return decoration;
+            })
     {
     }
 };
 
-miral::DecorationBasicManager::DecorationBasicManager(Decoration::DecorationBuilder decoration_builder)
-    : self{std::make_shared<Self>(decoration_builder)}
+miral::DecorationBasicManager::DecorationBasicManager(
+    mir::Server& server, Decoration::DecorationBuilder&& decoration_builder) :
+    self{std::make_shared<Self>(server, std::move(decoration_builder))}
 {
 }
